+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Bugfixes
+- Fixed rare crash when unable to load Mooshroom textures.
 - [FORGE] Fixed lockdown effects not syncing when swapping dimensions or joining worlds.